bascule de la classe AthleteView dans "models"
